Add performance tests for pulp deb
closes #970
