+++ .github/workflows/nightly.yml
@@ -117,3 +117,109 @@ jobs:
         run: |
           tar -xvf docs.tar -C ./docs
           .github/workflows/scripts/publish_docs.sh nightly ${GITHUB_REF##*/}
+
+  performance:
+    runs-on: ubuntu-latest
+    needs: test
+
+    strategy:
+      fail-fast: false
+      matrix:
+        env:
+          - TEST: performance
+            PERFORMANCE_TEST: sync
+          - TEST: performance
+            PERFORMANCE_TEST: publish
+          - TEST: performance
+            PERFORMANCE_TEST: pulp_to_pulp
+
+    steps:
+      - uses: "actions/checkout@v4"
+        with:
+          fetch-depth: 1
+          path: "pulp_deb"
+
+      - uses: "actions/checkout@v4"
+        with:
+          fetch-depth: 1
+          repository: "pulp/pulp-openapi-generator"
+          path: "pulp-openapi-generator"
+
+      - uses: actions/download-artifact@v3
+        with:
+          name: "plugin_package"
+          path: "pulp_deb/dist/"
+
+      - uses: "actions/setup-python@v4"
+        with:
+          python-version: "3.8"
+
+      - name: "Set environment variables"
+        run: |
+          echo "TEST=${{ matrix.env.TEST }}" >> $GITHUB_ENV
+          echo "PERFORMANCE_TEST=${{ matrix.env.PERFORMANCE_TEST }}" >> $GITHUB_ENV
+
+      - name: "Install python dependencies"
+        run: |
+          echo ::group::PYDEPS
+          pip install towncrier twine wheel httpie docker netaddr boto3 ansible mkdocs
+          echo "HTTPIE_CONFIG_DIR=$GITHUB_WORKSPACE/pulp_deb/.ci/assets/httpie/" >> $GITHUB_ENV
+          echo ::endgroup::
+
+      - name: "Before Install"
+        run: |
+          .github/workflows/scripts/before_install.sh
+        shell: "bash"
+        env:
+          PY_COLORS: "1"
+          ANSIBLE_FORCE_COLOR: "1"
+          GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
+          GITHUB_CONTEXT: "${{ github.event.pull_request.commits_url }}"
+
+      - name: "Install"
+        run: |
+          .github/workflows/scripts/install.sh
+        shell: "bash"
+        env:
+          PY_COLORS: "1"
+          ANSIBLE_FORCE_COLOR: "1"
+          GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
+          GITHUB_CONTEXT: "${{ github.event.pull_request.commits_url }}"
+
+      - name: "Install Python client"
+        run: |
+          .github/workflows/scripts/install_python_client.sh
+        shell: "bash"
+
+      - name: "Before Script"
+        run: |
+          .github/workflows/scripts/before_script.sh
+        shell: "bash"
+        env:
+          PY_COLORS: "1"
+          ANSIBLE_FORCE_COLOR: "1"
+          GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
+          GITHUB_CONTEXT: "${{ github.event.pull_request.commits_url }}"
+          REDIS_DISABLED: "${{ contains('', matrix.env.TEST) }}"
+
+      - name: "Performance Test"
+        run: |
+          .github/workflows/scripts/script.sh
+        shell: "bash"
+        env:
+          PY_COLORS: "1"
+          ANSIBLE_FORCE_COLOR: "1"
+          GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
+          GITHUB_CONTEXT: "${{ github.event.pull_request.commits_url }}"
+
+      - name: "Logs"
+        if: always()
+        run: |
+          echo "Need to debug? Please check: https://github.com/marketplace/actions/debugging-with-tmate"
+          http --timeout 30 --check-status --pretty format --print hb "https://pulp${PULP_API_ROOT}api/v3/status/" || true
+          docker images || true
+          docker ps -a || true
+          docker logs pulp || true
+          docker exec pulp ls -latr /etc/yum.repos.d/ || true
+          docker exec pulp cat /etc/yum.repos.d/* || true
+          docker exec pulp bash -c "pip3 list && pip3 install pipdeptree && pipdeptree"
