+++ CHANGES/970.misc
@@ -1,0 +1,1 @@
+Added sync, publish and pulp2pulp performance tests to run with the nightly CI.
